Move tests to SDK.Tests.
